ml/pyproject.toml
[project]
name = "aishwariya"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.14",
    "openai>=1.93.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "uvicorn>=0.35.0",
]
